+++ examples/js-example/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @stackframe/js-example
 
+## 2.8.8
+
+### Patch Changes
+
+- @stackframe/js@2.8.8
+
+## 2.8.7
+
+### Patch Changes
+
+- @stackframe/js@2.8.7
+
 ## 2.8.6
 
 ### Patch Changes