+++ examples/supabase/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @stackframe/example-supabase
 
+## 2.8.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack@2.8.8
+
+## 2.8.7
+
+### Patch Changes
+
+- @stackframe/stack@2.8.7
+
 ## 2.8.6
 
 ### Patch Changes