+++ packages/stack-ui/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @stackframe/stack-ui
 
+## 2.8.8
+
+### Patch Changes
+
+- Various changes
+  - @stackframe/stack-shared@2.8.8
+
+## 2.8.7
+
+### Patch Changes
+
+- @stackframe/stack-shared@2.8.7
+
 ## 2.8.6
 
 ### Patch Changes